:test_tube: [ADD] Scroll bar porcent logic encapsulated in a custom hook
Test for progressBar component
