+++ package.json
@@ -24,6 +24,7 @@
 	"devDependencies": {
 		"@testing-library/jest-dom": "^5.14.1",
 		"@testing-library/react": "^12.0.0",
+		"@testing-library/react-hooks": "^7.0.1",
 		"@testing-library/user-event": "^13.2.1",
 		"@types/jest": "^26.0.24",
 		"@types/node": "^16.4.0",
