+++ styles/styledComponents/ProgressBar.ts
@@ -13,8 +13,6 @@ export const ProgressScrollStyled = styled.section<props>`
 	z-index: 5;
 
 	.progress__hidden {
-		position: fixed;
-		top: 0;
 		width: 100%;
 		height: 10px;
 		background: linear-gradient(
@@ -22,7 +20,6 @@ export const ProgressScrollStyled = styled.section<props>`
 			transparent ${p => p.scroll}%,
 			${p => p.theme.background} 0%
 		);
-		transition: background 1s;
 		z-index: 10;
 	}
 `;
